refactor: move helm nginx helpers to the dedicated file
For better modularity and readability.
