+++ core/src/plugins/kubernetes/local/k3s.ts
@@ -12,7 +12,7 @@ import { KubeApi } from "../api.js"
 import type { KubernetesResource } from "../types.js"
 import type { PluginContext } from "../../../plugin-context.js"
 import type { SystemVars } from "../init.js"
-import type { NginxHelmValuesGetter } from "../nginx/ingress-controller.js"
+import type { NginxHelmValuesGetter } from "../nginx/nginx-helm.js"
 
 export async function isK3sFamilyCluster(ctx: PluginContext, provider: KubernetesProvider, log: Log): Promise<boolean> {
   return await isK3sFamilyClusterContext(ctx, provider, log)
