+++ core/src/plugins/kubernetes/nginx/ingress-controller.ts
@@ -20,12 +20,7 @@ import { apply } from "../kubectl.js"
 import { minikubeNginxInstall, minikubeNginxStatus, minikubeNginxUninstall } from "../local/minikube.js"
 import { defaultBackendInstall, defaultBackendStatus, defaultBackendUninstall } from "./default-backend.js"
 import { getK3sNginxHelmValues } from "../local/k3s.js"
-
-const HELM_INGRESS_NGINX_REPO = "https://kubernetes.github.io/ingress-nginx"
-const HELM_INGRESS_NGINX_VERSION = "4.0.13"
-const HELM_INGRESS_NGINX_CHART = "ingress-nginx"
-const HELM_INGRESS_NGINX_RELEASE_NAME = "garden-nginx"
-const HELM_INGRESS_NGINX_DEPLOYMENT_TIMEOUT = "300s"
+import { getGenericNginxHelmValues, helmNginxInstall, helmNginxStatus, helmNginxUninstall } from "./nginx-helm.js"
 
 export async function ingressControllerReady(ctx: KubernetesPluginContext, log: Log): Promise<boolean> {
   const clusterType = ctx.provider.config.clusterType
@@ -105,134 +100,4 @@ export async function ingressClassCreate(ctx: KubernetesPluginContext, log: Log)
     },
   }
   await apply({ log, ctx, api, provider, manifests: [ingressClass], namespace })
-}
-
-export async function helmNginxStatus(ctx: KubernetesPluginContext, log: Log): Promise<DeployState> {
-  const provider = ctx.provider
-  const config = provider.config
-
-  const namespace = config.gardenSystemNamespace
-  try {
-    const statusRes = JSON.parse(
-      await helm({
-        ctx,
-        log,
-        namespace,
-        args: ["status", HELM_INGRESS_NGINX_RELEASE_NAME, "--output", "json"],
-        // do not send JSON output to Garden Cloud or CLI verbose log
-        emitLogEvents: false,
-      })
-    )
-    const status = statusRes.info?.status || "unknown"
-    log.debug(chalk.yellow(`Helm release status for ${HELM_INGRESS_NGINX_RELEASE_NAME}: ${status}`))
-    return helmStatusMap[status] || "unknown"
-  } catch (error) {
-    log.debug(chalk.yellow(`Helm release ${HELM_INGRESS_NGINX_RELEASE_NAME} missing.`))
-    return "missing"
-  }
-}
-
-// TODO: consider using some specific return type here, maybe something from helm SDK?
-export type NginxHelmValuesGetter = (systemVars: SystemVars) => object
-
-export const getGenericNginxHelmValues: NginxHelmValuesGetter = (systemVars: SystemVars) => {
-  return {
-    name: "ingress-controller",
-    controller: {
-      kind: "DaemonSet",
-      updateStrategy: {
-        type: "RollingUpdate",
-        rollingUpdate: {
-          maxUnavailable: 1,
-        },
-      },
-      extraArgs: {
-        "default-backend-service": `${systemVars.namespace}/default-backend`,
-      },
-      hostPort: {
-        enabled: true,
-        ports: {
-          http: systemVars["ingress-http-port"],
-          https: systemVars["ingress-https-port"],
-        },
-      },
-      minReadySeconds: 1,
-      tolerations: systemVars["system-tolerations"],
-      nodeSelector: systemVars["system-node-selector"],
-      admissionWebhooks: {
-        enabled: false,
-      },
-      ingressClassResource: {
-        name: "nginx",
-        enabled: true,
-        default: true,
-      },
-    },
-    defaultBackend: {
-      enabled: false,
-    },
-  }
-}
-
-export async function helmNginxInstall(
-  ctx: KubernetesPluginContext,
-  log: Log,
-  nginxHelmValuesGetter: NginxHelmValuesGetter
-) {
-  const nginxStatus = await helmNginxStatus(ctx, log)
-  const backendStatus = await defaultBackendStatus(ctx, log)
-
-  if (nginxStatus === "ready" && backendStatus === "ready") {
-    return
-  }
-
-  const provider = ctx.provider
-  const config = provider.config
-  const namespace = config.gardenSystemNamespace
-  const systemVars: SystemVars = getKubernetesSystemVariables(config)
-  const values = nginxHelmValuesGetter(systemVars)
-
-  const valueArgs: string[] = []
-  for (const key in values) {
-    if (values.hasOwnProperty(key)) {
-      valueArgs.push(`${key}=${JSON.stringify(values[key])}`)
-    }
-  }
-
-  // TODO-G2: update the nginx version
-  const args = [
-    "install",
-    HELM_INGRESS_NGINX_RELEASE_NAME,
-    HELM_INGRESS_NGINX_CHART,
-    "--version",
-    HELM_INGRESS_NGINX_VERSION,
-    "--repo",
-    HELM_INGRESS_NGINX_REPO,
-    "--timeout",
-    HELM_INGRESS_NGINX_DEPLOYMENT_TIMEOUT,
-    "--set-json",
-    `${valueArgs.join(",")}`,
-  ]
-
-  log.info(`Installing nginx in ${namespace} namespace...`)
-  await defaultBackendInstall(ctx, log)
-  await helm({ ctx, namespace, log, args, emitLogEvents: false })
-
-  log.success(`nginx successfully installed in ${namespace} namespace`)
-}
-
-export async function helmNginxUninstall(ctx: KubernetesPluginContext, log: Log) {
-  // uninstall default-backend first
-  await defaultBackendUninstall(ctx, log)
-
-  const status = await helmNginxStatus(ctx, log)
-  if (status === "missing") {
-    return
-  }
-
-  const provider = ctx.provider
-  const config = provider.config
-  const namespace = config.gardenSystemNamespace
-
-  await helm({ ctx, namespace, log, args: ["uninstall", HELM_INGRESS_NGINX_RELEASE_NAME], emitLogEvents: false })
 }
